fix: refetch after verified members motion finalized
